+++ src/Detach/Collisions/Primitives3D/Triangle3D.cs
@@ -2,13 +2,13 @@
 
 namespace Detach.Collisions.Primitives3D;
 
-public record struct Triangle
+public record struct Triangle3D
 {
 	public Vector3 A;
 	public Vector3 B;
 	public Vector3 C;
 
-	public Triangle(Vector3 a, Vector3 b, Vector3 c)
+	public Triangle3D(Vector3 a, Vector3 b, Vector3 c)
 	{
 		A = a;
 		B = b;