+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Moved 2D and 3D primitives to `Detach.Collisions.Primitives2D` and `Detach.Collisions.Primitives3D` namespaces respectively.
+- Renamed `Triangle` to `Triangle3D`.
 
 ## 0.9.1